Fix integer overflow in find_address_in_search_tree bounds check
The addition of node_count and data_section_size was performed in
uint32_t arithmetic. On very large databases where the sum exceeds
UINT32_MAX, this wraps and causes valid records to be rejected as
corrupt. Cast to uint64_t before addition.

+++ src/maxminddb.c
@@ -985,7 +985,7 @@ static int find_address_in_search_tree(const MMDB_s *const mmdb,
 
     result->netmask = current_bit;
 
-    if (value >= node_count + mmdb->data_section_size) {
+    if (value >= (uint64_t)node_count + mmdb->data_section_size) {
         // The pointer points off the end of the database.
         return MMDB_CORRUPT_SEARCH_TREE_ERROR;
     }
